moved LoginUseCase to oauth module
